requirements.txt
# Создание файлов с данными
Faker==22.6.0
unidecode==1.3.6

# Базы данных
pymongo>=4.5.0,<5.0.0
psycopg2-binary>=2.9.9,<3.0.0
apache-airflow-providers-mongo

# Kafka
kafka-python==2.0.2
apache-airflow-providers-apache-kafka

# Для работы с S3
boto3>=1.26.0
s3fs>=2023.6.0

# Spark
pyspark>=3.4.2
py4j==0.10.9.7
apache-airflow-providers-apache-spark

# Для ClickHouse
clickhouse-driver>=0.2.9
airflow-clickhouse-plugin
clickhouse-connect>=0.8.0
airflow-providers-clickhouse==0.0.1

# HTTP requests
requests

# Утилиты
python-dotenv==1.0.1




